examples: exit early with clean message for examples that require more than 2 outputs

+++ src/audio_worklet_node.rs
@@ -306,7 +306,7 @@ fn process_audio_worklet(
     // fill AudioWorkletGlobalScope
     let mut global = env.get_global()?;
     global.set_named_property("currentTime", current_time)?;
-    global.set_named_property("currentFrame", current_frame)?;
+    global.set_named_property("currentFrame", current_frame as u64)?;
 
     let k_worklet_callable_process =
         env.symbol_for("node-web-audio-api:worklet-callable-process")?;
